requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.28.1
requests==2.31.0
pymongo==4.6.2
pydantic==2.6.4
tqdm==4.66.2
python-multipart==0.0.9
pillow==10.2.0
dlib==19.24.2
deepface==0.0.89
tf-keras==2.16.0
Unidecode~=1.3.8
pandas~=2.2.1
matplotlib~=3.8.4
